[ADD][MAJOR] Implementation and cut of function
This is a big commit and as such will have a big description.
In this commmit the files that use to be for detection have been cut.
The video flux detection has been fused due to beeing redondant.
The picture detection has been move and changed.
All those ffiles have been moved and renamed.
The main file has been updated to reflect those changes.